requirements.txt
asgiref==3.7.2
Django==4.2.5
django-admin-cli==0.1.1
django-cors-headers==4.2.0
django-filter==23.3
djangorestframework==3.14.0
Markdown==3.4.4
psycopg2-binary==2.9.8
python-dotenv==1.0.0
pytz==2023.3.post1
sqlparse==0.4.4
